feat: combine trade ix to opening position
